% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_outcome.R
\docType{data}
\name{cont_outcome}
\alias{cont_outcome}
\title{Studies Comparing Truncal Blocks on Pain Outcomes}
\format{
A data frame with 12 rows and 15 variables:
\describe{
\item{Author}{Study author}
\item{Year}{Year study published}
\item{Subgroup}{Intervention subgroup}
\item{Mean_Control}{Control group pain score mean}
\item{SD_Control}{Control group pain score standard deviation}
\item{N_Control}{Number of patients in the control group}
\item{Mean_Intervention}{Intervention group pain score mean}
\item{SD_Intervention}{Intervention group pain score standard deviation}
\item{N_Intervention}{Number of patients in the intervention group}
\item{D1}{Domain 1 of the Risk of Bias 2.0 tool}
\item{D2}{Domain 2 of the Risk of Bias 2.0 tool}
\item{D3}{Domain 3 of the Risk of Bias 2.0 tool}
\item{D4}{Domain 4 of the Risk of Bias 2.0 tool}
\item{D5}{Domain 5 of the Risk of Bias 2.0 tool}
\item{Overall}{Overall risk of bias}
}
}
\usage{
cont_outcome
}
\description{
Outcome data from randomised controlled trials comparing truncal blocks (eg TAP blocks, rectus sheath blocks, perperitoneal blocks) against comparator analgesia or placebo in pain outcomes for patients undergoing midline laparotomy.
}
\keyword{datasets}
